cr50: enable UART0RX when external pull up is detected
Disabling UART0 RX on the development board prevents us from working
with it easily. This change moves UART0 RX to where it is on the
platforms and will enable the UART RX if it sees that DIOA13 is
externally pulled high. On development boards, you can connect DIOA1 to
DIOA13 to use the Cr50 console.

BUG=none
BRANCH=none
TEST=On the development board test that cr50 boots fine and that the
debug console is read only. Connect a jumper between DIOA1 and DIOA13
and check that the console is now read write. Check that Cr50 on reef
and Gru still boot normally.

Change-Id: Id4c9bb578b7d142ae3970fa9501a818af08a355e