Update patch set 1
Patch Set 1:

they still would have to go in different patches. using 'response' in this sense in this file would confuse the hell out of any reviewer.

The server side is not nearly as confusuing, you can fix it later, but let's keep this patch - I have some more on top of it and it makes code much easier to understand.

Patch-set: 1