pd: FUSB302B: WIP - Modified driver to operate in manual detect mode
The existing driver code handles SRC/SNK mode differently for
it's attach logic. This is due to an issue of Rev A silicon that
didn't allow CC1 & CC2 lines to pulled up independently. The
workaround for that issue is to enable the FUSB302's autodetect DRP
circuit.

On Kevin when a SNK device is plugged in, the correct CC line is
often misidentified and the result is that the PD state machine
gets stuck in SRC_DISCOVERY.

This CL disables the autodetect mode for SRC and appears to correct
the problem with getting stuck in SRC_DISCOVERY mode.

BUG=chrome-os-partner:54790
BRANCH=none
TEST=manual

Tested with ANKER TypeC hub. Verified that it correctly attaches
where before it often was stuck in SRC_DISCOVERY mode. Also
verified that the same CC line is selected when the same physical
orientation is used each connect attempt.

Note there is still an issue connecting either HoHo or DingDong adapters.

Change-Id: I9b0e23e65309948f3398a9fd1f072b2c3070a306
Signed-off-by: Scott <scollyer@chromium.org>