CR50: match privkey against certs on endorsement
This patch adds code for generating RSA and ECC keys
from the endorsement seed.  The generated keys are
matched against the certificates being installed.

The code for key generation is time consuming and
slows down certificate installation.  Hence this
part of the verification step should be removed
from a future version of the code.

Certificates are not fully parsed, and hence
a cert that contains the corresponding public key
is considered to be a match.

Also filed CRBUG/P/55260, in which the TPMT_PUBLIC
template is truncated during personalization,
and hence requires a workaround in RW firmware.

BRANCH=none
BUG=chrome-os-partner:43025,chrome-os-partner:47524
BUG=chrome-os-partner:50115,chrome-os-partner:55260
TEST=test full personalize + cros_ack verify cert flow

Change-Id: I1af83f1ec86e7ee4d325a4b7aabe03ce08c4108b
Signed-off-by: nagendra modadugu <ngm@google.com>