elm: allow charging of dead battery requesting nil
On elm battery, when the battery is dead it reports 0 for desired
voltage, current, and state of charge. In this case we should allow
charging.

Added a CONFIG option for this that should be removed as soon as
the battery side is fixed.

With this CL, when a dead oak battery is used and a charger is
connected, we attempt to charge it.

BUG=chrome-os-partner:54944
BRANCH=none
TEST=test on an elm with a dead battery. w/o this CL, the battery
never charges because the charging not allowed flag is set. With this
CL, the battery charges.

Change-Id: I90e2407a265710eeb3d175bdb0572055cc939afd
Signed-off-by: YH Huang <yh.huang@mediatek.com>