Update patch set 4
Patch Set 4: Verified-1

This patch results in incorrect orientation.

The accelerometer orientation frame should be as follows:

                    +
                    |
                    |
                    | Lid
                    |
                    |
         +----------+
             Base
+-------------------------------+

Lid:
 - The positive Z axis goes out of the screen towards the left of the page.
 - The positive Y axis goes into the page.
 - The positive X axis goes up towards the top of the page.

Base:
 - The positive Z axis goes into the base towards the bottom of the page.
 - The positive Y axis goes into the page.
 - The positive X axis goes towards the left of the page.

This is such that when the device is closed in normal mode, the axes align.

Below are my expectations of the sign and magnitude.

Device is sitting on table with lid open at 90 degrees.
Base: 0, 0, 1
Lid: -1, 0, 0

Device is rotated to rest on its left side.
Base: 0, 1, 0
Lid:  0, 1, 0

Device is resting on its lid on the table.
Base: -1, 0, 0
Lid:  0, 0, -1

I tested without this change, and the orientations are correct but the signs are wrong. But with this change, the orientations and signs are wrong.

Patch-set: 4
Label: Verified=-1