Update patch set 4
Patch Set 4:

Sorry, my last comment was incorrect. I just realized that I was looking at an older doc. The accelerometer orientation should be the following:

"Accelerometer is located in the device's screen. In the screen's natural
orientation, positive X points to the right, consistent with the pixel
position. Positive Y points up the screen. Positive Z is perpendicular to
the screen, pointing outwards towards the user. The orientation is
described at:
 http://www.html5rocks.com/en/tutorials/device/orientation/.

Accelerometer is located in a keyboard attached to the device's screen.
If the device is open 180 degrees the orientation is consistent with the
screen. I.e. Positive X points to the right, positive Y points up on the
keyboard and positive Z is perpendicular to the keyboard pointing out
towards the user."

Examining the values again, the lid seems correct. But the base needs some changes.

Patch-set: 4