anx74xx: Do not define ANX74XX_INT_ACTIVE_POLARITY
Some boards need to set this to low. We could have some messy #undef
and #ifndef logic to change the default, but IMO it's better just to
get rid of this default since it's board-specific and the compiler
should complain if the developer forgets to set it correctly.

BUG=none
BRANCH=none
TEST=builds for boards with this set.

Change-Id: I06a58c4e26892843243e8e98f2c86c6d3a696eb1
Signed-off-by: David Hendricks <dhendrix@chromium.org>