CR50: do not try searching in uninitialized TPM NV RAM.
The manufacturing status check verifies if the proper certificates are
found in the device NV RAM. This check can not succeed unless NV RAM
metadata is initialized by calling _TPM_Init().

If the check shows that the device has not been through manufacturing
sequence yet, TPM_Manufacture() needs to be invoked to make sure that
all relevant TPM structures are initialized and properly stored in NV
RAM. _TPM_Init() needs to be invoked again after that.

BRANCH=ToT
BUG=chrome-os-partner:43025
TEST=restarting Kevin device with pre-manufactured CR50 takes it
     through factory initialization on every reboot. Restarting Kevin
     once TPM is through manufacturing process shows that the
     previously saved rollback counters are preserved.

Change-Id: I80b69f2c4b8d0e4cca154db510867df39c707ce2
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>