Fix unit of MSDN time interface clock() to ms
The unit of the MSDN clock() interface is milliseconds,
while the current implementation return microseconds.

This change fixes the units.  The only callers of
this function are the TPM implementation on CR50,
and test code.

The TPM implementation expects milliseconds, so no
change there.  test/nvmem.c has been updated to
call get_time() in place of clock().

BRANCH=none
BUG=chrome-os-partner:43025,chrome-os-partner:47524
BUG=chrome-os-partner:50115
TEST=all lockout related TCG tests pass

Change-Id: Ic02fffca610426d22e58609eb8c3693aec96ad5c
Signed-off-by: nagendra modadugu <ngm@google.com>