Update patch set 1
Patch Set 1:

I don't think this is sufficient/guaranteed to be correct.

does tpm_init() have any vars in .data which in this case are not going to be reset to the original values?

Are there any variables in .bss which are not being set to zero?

Patch-set: 1
Label: Trybot-Ready=0