tcpm: fusb302: Remove bc_lvl interrupt enable and status read
The bc_lvl interrupt was enabled in the init module and
its status checked for every FUSB302 interrupt indication
to the EC. This interrupt then sets PD_EVENT_CC event.
However, this event is only checked if the #define
is set.

Having this interrupt continuously enabled is troublesome
when PD protocol messages are being sent/received. The CC\
line is changing states by design during these times and
this results in uncessary interrupts and PD task delays.

Having this interrupt enabled causes a problem when attaching
either ding dong or hoho display adapters. Removing the bc_lvl
interrupt eliminates the problem and as stated above, the the
interrupt serves no purpose to the USB PD protocol state machine
that's running on the EC.

BUG=chrome-os-partner:54790
BRANCH=none
TEST=manual

Change-Id: I4b5c20740d35722fc20d7307d0c3d1cd317964ee
Signed-off-by: Scott <scollyer@chromium.org>