Update patch set 2
Patch Set 2:

No, there are two of them. When I removed the intervening CL to undef CONFIG_FLASH, gerrit created a new Change-Id. That's the one that was abandoned - it looks exactly the same except for that. I put the original Change-Id back and uploaded again so it would show up as patch 2, instead of a new CL.

Patch-set: 2