NOT FOR MERGING: one off patch for migrating to new ro
Recent remapping of the flash memory layout and introduction of the
new RO image supporting RW update fallbacks introduces a problem of
migrating the currently used CR50 chips to the new flash memory
layout/RO image.

This patch does exactly that: early on startup it checks if the RO-B
space contains a valid RO image. If it does not, the code presumes
that it is running on a device with the old flash layout and starts
migration.

First the new NVRAM location is erased and the existing NVRAM contents
is moved to the new location. NVRAM contents is position independent
and can be used anywhere.

The RO-B images for both B1 and B2 versions of the chip are included.
The code determines the SOC version it is running on and copies the
appropriate RO-B image into the proper location.

Both NVRAM relocation and RO-B image update are verified, and if
everything is fine, RO-A is corrupted and the device is rebooted,
coming back up with the new RO and flash memory layout.

BRANCH=none
BUG=none

TEST=tried this on both b1 and b2 dev boards. Discovered that once the
     new RO is installed, RW updates stop working; the update is
     accepted, but then after reset the new RO complaints that there
     is no valid RW image present.

     Not sure what the problem is, but all signs show at some kind of
     RO_B bug.

Change-Id: I398fdd25a2bde711be94155f680455881eaf9eb8
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>