Update patch set 12
Patch Set 12:

(1 comment)

> > (1 comment)
 > 
 > https://chromium-review.googlesource.com/#/c/361250/
 > 
 > Also, why is this change in firmware branch only? Aren't we
 > supposed to commit to master branch first?

This is more like a HACK trying to reset the bad sensors. We don't know the root cause yet, and we don't even know if this works or not since we cannot reproduce it. It is better to make this stay in firmware branch until we fully understand what is the root cause.

Patch-set: 12