WIP: Generate a version string that includes all source repos
Update the getversions.sh script to generate a combined version string for
all repos that contribute source code to the build.

It should be possible to find 3/4 of the repos using CROS_WORKON_DESTDIR,
but the private-cr51 directory is not mentioned in that variable.

The generation of a combined string is controlled by an environment
variable called EXTENDED_VERSION. If this is set (e.g. to "yes") in the
ebuild, then the extended version will be used.

The EC version code is updated to use the combined string in version.c.
However the USB code (board/cr50/board.c has not been changed). Most
likely the 32-character version of the combined string is not useful.

Once I have a bit of feedback on what people think is the best approach I
can fiddle a little more.

Signed-off-by: Simon Glass <sjg@chromium.org>
BUG=chrome-os-partner:55373
BRANCH=none
TEST=emerge-gru chromeos-ec; strings /build/gru/firmware/ec.bin |grep gru_;
See a combined version string.
NOTE this has not been tested on cr50 or with 'version' command!
Change-Id: I2c211cbda8c3cab3c8c21b4430e4b3102691e74a