Update patch set 2
Patch Set 2:

(1 comment)

Addressed shawnn@ comment and also modified the detect_cc_pin_source_manual() function so that it won't measure a cc line that has VCONN active.

I will plan on removing support for rev A, but will do that in a subsequent CL so that the changes are easier to track.

Patch-set: 2
Label: Code-Review=0