Update patch set 2
Patch Set 2:

I actually think we can drop some of this to free up space in the EC.

DPTF on the host never managed to enable support for the threshold triggers, so they are unused on all platforms.  And due to the new mostly userspace architecture for DPTF I don't think it will ever be brought back.

We still need the thermal sensors to be available in the ACPI memory map however.

Also, enabling DPTF does not necessarily imply that the EC is not doing thermal control.  There are some platforms where DPTF is used for passive cooling but the fans are controlled by the EC.

Patch-set: 2