g: do not print unrelated version information.
The recent addition of the multicomponent version string for cr50,
requires further tweaking of the version reporting routines. In cases
where version RW_A or RW_B needs to be reported, it is enough to
include just the first word of the combined version string, i.e.
everything til the first white space character.

It is also possible to reduce the size of memory allocated for storing
version strings, as there is never more than one string required at a
time.

BRANCH=none
BUG=chrome-os-partner:55373
TEST=verified that RO_x and RW_x versions are printed properly:
  > vers
  Chip:    g cr50 B2
  Board:   0
  RO_A:    0.0.1/84e2dde7
  RO_B:  * 0.0.2/13eda43f
  RW_A:  * cr50_v1.1.4979-0061603+
  RW_B:    cr50_v1.1.4978-72cb7e3+
  Build:   cr50_v1.1.4979-0061603+
  ...

Change-Id: I192eb29816dfa963b08aa97f749b978b1367d6b7
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>