g: fix rdd
It looks like when the EC is off and we change the value of the debug
map, another RDD interrupt is triggered. Then we would check the cc1 and
cc2 values and detect the cable is connected then rewrite the debug map
value again. This would trigger another interrupt and continue the cycle
until the watchdog stepped in at which point cr50 would reset and the EC
would be up by then so it would detect the debug cable and move on.

This change adds a check to see that the RDD was trying to detect the
state it currently senses before doing anything so the second interrupt
will be ignored.

There is another issue when after a soft reset the debug map value does
not reset to default. It remains the value it was before the soft reset.
This messed up the debug connected disconnected sense values.

This change defines the debug detect value and uses that to populate the
debug map on init.

BUG=chrome-os-partner:55793
BRANCH=none
TEST=manual
	plug in a reworked suzyq

	Flash the EC

	verify a bunch of RDD interrupts aren't triggered and the
	reworked suzyq can be used.

	plug it in and out a couple of times to make sure everything
	works.

Change-Id: I6f4e7b15cf9c53b965533df7fe23fae5b340f70d
Signed-off-by: Mary Ruthven <mruthven@chromium.org>