Update patch set 1
Patch Set 1: Code-Review+1

(1 comment)

I won't claim that I fully understand this patch, but it has a couple of very good side effects when applied to a cr50 on Kevin:

With a recent addition of the reboot counter, I observed that the counter was always 3 after an update or 2 after a restart, which indicated there was an extra reset while the system was booting.

Also, for some reason the very first access to the TPM to read vid:did was failing while suzy-q is connected, I was in the process of investigating this phenomenon.

With this patch applied both problems are gone, great job!

Patch-set: 1
Label: Code-Review=+1