Update patch set 1
Patch Set 1:

So the way this works is that interrupt and reset are separated in time by the watchdog controller.  When the count first expires, the interrupt fires.  The count internally is then reloaded.  IF the count expires again, and the watchdog unit sees the reset STILL has not been handled, then it requests for reset.  So I think this is pretty ideal behavior right? On the first one, firmware should attempt to handle it, if for some reason interrupts are wedged, then the hammer comes down later.

Patch-set: 1