Update patch set 1
Patch Set 1:

If we enable this, it will cause certain commands (EC_CMD_FLASH_ERASE) to return EC_RES_IN_PROGRESS immediately while the host command task continues to process the command (eg. continues to erase a large block of flash).

Do we know what was originally sending this command? I wouldn't expect it to be sent by the host unless a command returns status EC_RES_IN_PROGRESS.

Why do we typically enable this for ARM platforms and not x86? Is it because polling host command completion status (which the host will do constantly while a command is pending) is super light in LPC vs SPI / I2C?

Patch-set: 1